Moved relevant files into new package
